+++ WebStoreAPI/Commands/Users/DeleteUserHandler.cs
@@ -20,6 +20,11 @@ namespace WebStoreAPI.Commands.Users
         {
             var user = await _context.Users.FirstOrDefaultAsync(x => x.Id == command.Id, cancellationToken);
 
+            if (user == null)
+            {
+                return new Unit();
+            }
+
             _context.Users.Remove(user);
             await _context.SaveChangesAsync(cancellationToken);
             return Unit.Value;
